fix: (HDS-2651) only latest version of documentation for development mode
