+++ site/gatsby-node.js
@@ -1,5 +1,6 @@
 const webpack = require('webpack');
 const path = require('path');
+const isDevelopmentMode = process.env.NODE_ENV === 'development';
 
 exports.onCreateWebpackConfig = ({ actions, getConfig }) => {
   const config = getConfig();
@@ -59,6 +60,24 @@ exports.createPages = async ({ actions, graphql }) => {
   // GraphQL query to fetch frontmatter and rendered MDX content
   const result = await graphql(`
     query SiteDataQuery {
+      site {
+        siteMetadata {
+          title
+          description
+          siteUrl
+          menuLinks {
+            name
+            link
+            subMenuLinks {
+              name
+              link
+              withDivider
+            }
+          }
+          footerTitle
+          footerAriaLabel
+        }
+      }
       allMdx {
         edges {
           node {
@@ -72,9 +91,10 @@ exports.createPages = async ({ actions, graphql }) => {
             parent {
               ... on File {
                 relativePath
-                gitRemote {
-                  ref
-                }
+                ${!isDevelopmentMode ?
+                `   gitRemote {
+                      ref
+                }` : ``}
               }
             }
           }
@@ -87,6 +107,76 @@ exports.createPages = async ({ actions, graphql }) => {
     console.error(result.errors);
     throw new Error('Failed to fetch MDX data');
   }
+
+  const splitPathIntoParts = (path) => path.split('/').filter((l) => !!l);
+
+  const isLinkParentForPage = (parentPath, level) => (page) => {
+    const pathParts = splitPathIntoParts(page.slug);
+    const ret = pathParts.length === level && pathParts.slice(0, -1).every((pathPart) => parentPath.includes(pathPart));
+    return ret;
+  };
+
+  const sortByPageTitle = (pageA, pageB) => pageA.title.localeCompare(pageB.title);
+
+  const generateUiIdFromPath = (path, prefix) => {
+    const pathStr =
+      !path && path === '/'
+        ? 'home'
+        : path
+            .split('/')
+            .filter((str) => !!str)
+            .join('-');
+    return `${prefix}-${pathStr}`;
+  };
+
+  const resolveCurrentMenuItem = (version, menuItems, slugWithPrefix) => {
+    const rootPath = '/';
+
+    if (slugWithPrefix === rootPath) {
+      return menuItems.find(({ link }) => link === rootPath);
+    } else {
+      const ret = menuItems
+        .filter(({ link }) => link !== rootPath)
+        .find((menuItem) => slugWithPrefix.startsWith(menuItem.link));
+      return ret;
+    }
+  };
+
+  const getUiSubMenuLinks = (allPages, uiMenuLinks, currentMenuItem) => {
+    const subMenuLinks = currentMenuItem?.subMenuLinks || [];
+    const subMenuLinksFromPages =
+      currentMenuItem && currentMenuItem.link
+        ? allPages
+            .filter(isLinkParentForPage(currentMenuItem.link, 2))
+            .map((page) => ({ name: page.title, title: page.title, link: page.slug }))
+            .sort(sortByPageTitle)
+        : [];
+
+    const uiSubMenuLinks = [...subMenuLinks, ...subMenuLinksFromPages].map((subMenuLink) => ({
+      ...subMenuLink,
+      prefixedLink: subMenuLink.link,
+      uiId: generateUiIdFromPath(subMenuLink.link, 'side-nav'),
+      subLevels: allPages
+        .filter(isLinkParentForPage(subMenuLink.link, 3))
+        .map((subLevelLink) => ({
+          ...subLevelLink,
+          uiId: generateUiIdFromPath(subLevelLink.slug, 'side-nav-sub'),
+          prefixedLink: subLevelLink.slug,
+        }))
+        .sort(sortByPageTitle),
+    }));
+
+    return uiSubMenuLinks;
+  }
+
+  const siteData = result.data.site.siteMetadata;
+  const mdxPageData = result.data.allMdx?.edges || [];
+  const menuLinks = siteData?.menuLinks || [];
+
+  const uiMenuLinks = menuLinks.map((menuLink) => ({
+    ...menuLink,
+    uiId: generateUiIdFromPath(menuLink.link, 'nav'),
+  }));
 
   // Create pages dynamically
   result.data.allMdx.edges.forEach(({ node }) => {
@@ -103,12 +193,29 @@ exports.createPages = async ({ actions, graphql }) => {
         ? require.resolve(`./.cache/gatsby-source-git/docs-${gitRemote}/${node.parent.relativePath}`)
         : require.resolve(contentPath);
 
+      // filter out duplicate slug entries.
+      const allPages = Object.values(
+        Object.fromEntries(
+          mdxPageData
+            .filter(({ node }) => node?.parent?.gitRemote?.ref === gitRemote)
+            .filter(({ node }) => node.frontmatter.slug && node.frontmatter.navTitle)
+            .map(({ node }) => [node.frontmatter.slug, { ...node.frontmatter, ...node.fields }])
+        ),
+      );
+
+      const currentMenuItem = resolveCurrentMenuItem(gitRemote, uiMenuLinks, node.frontmatter.slug);
+      const uiSubMenuLinks = getUiSubMenuLinks(allPages, uiMenuLinks, currentMenuItem);
+
       createPage({
         component: `${pageTemplate}?__contentFilePath=${pageContent}`,
         path: pathWithVersion,
         context: {
           id: node.id,
           frontmatter: { ...node.frontmatter, slug: pathWithVersion },
+          siteData,
+          currentMenuItem,
+          uiMenuLinks,
+          uiSubMenuLinks
         },
       });
     } catch (e) {
