+++ site/src/data/versionsFromGit.json
@@ -1,5 +1,4 @@
 [
-  "4.2.0",
   "3.12.0",
   "2.17.1"
 ]
